Move torch compile util function to `utils.py`
